refactor: make header tsx
