+++ src/components/header.tsx
@@ -1,0 +1,21 @@
+import { Link } from "gatsby"
+import PropTypes from "prop-types"
+import React from "react"
+
+const Header = ({ siteTitle }) => (
+  <header>
+    <div className="flex bg-purple-400">
+      <h1 className="text-4xl ml-auto mr-6 my-4 font-bold">Bikes & Bytes</h1>
+    </div>
+  </header>
+)
+
+Header.propTypes = {
+  siteTitle: PropTypes.string,
+}
+
+Header.defaultProps = {
+  siteTitle: ``,
+}
+
+export default Header
